+++ dango/dex/src/lib.rs
@@ -1,6 +1,7 @@
 mod core;
+mod cron;
 mod execute;
 mod query;
 mod state;
 
-pub use {core::*, execute::*, query::*, state::*};
+pub use {core::*, cron::*, execute::*, query::*, state::*};
